Add ability to load a specific PSL section
The `PublicSuffixListManager` class is updated to enable its `getList` method
to return a `PublicSuffixList` object containing a specific section of the PSL

To avoid BC break, the following constants are added to the class:

- `PublicSuffixListManager::ICANN_SECTION` to load the ICANN section
- `PublicSuffixListManager::ICANN_PSL_PHP_FILE` to cache the ICANN section
- `PublicSuffixListManager::PRIVATE_SECTION` to load the PRIVATE section
- `PublicSuffixListManager::PRIVATE_PSL_PHP_FILE` to cache the PRIVATE section

An optional parameter is added to `PublicSuffixListManager::getList` to enable
loading a specific section. To avoild BC break if an invalid section name is
provided the method will return the full list.

The method `PublicSuffixListManager::refreshPublicSuffixList` is updated to
manage creating and updating each section cache list.