Introduce Rules::getPublicSuffix
Because Rules::resolve try to comply to PublicSuffix List algorithm some
exception are not always caught

Rules::getPublicSuffix is more strict than Rules::resolve and only returns the
public suffix information for a given domain.

The Domain and PublicSuffix value object now implements two modifiers
::toUnicode and ::toAscii to convers their value according to IDN UTS 46
algorithm. Domain names are still not validated but if the conversion can
not be perform an Exception is thrown.